requirements.txt
pydantic
python-dotenv
requests
PyMuPDF